Fix IH-setup (#30)
IH setup is failing with the error

Warning: No available formula with the name "rancher" (dependency of
consultingmd/ih-public/ih-core). Did you mean ranger? ==> Searching for
similarly named formulae and casks... ==> Formulae
rancher-cli ✔ rancher-compose ranger

To install rancher-cli ✔, run:
  brew install rancher-cli ✔
ih-setup is not available in your PATH. Check the logs above to see if
there was a fatal error, and make sure that /usr/local/bin is in your
PATH