[DT-142] Pre-commit handler
This adds a new, custom pre-commit handler (ih-pre-commit) which resides in the image-builder repository. This change updates the git step to remove its installation of the standard pre-commit hook, and updates the toolrepos step to install the ih-pre-commit step as a template and to update all existing IH repositories. It also introduces a shell script which can be run to pull the latest HEAD of a repo, or clone the repo if it is not present. This is used by the pre-commit hook to keep itself up-to-date.