DT-151 configure settings in rancher (#45)
In order to configure Rancher Desktop to disable kubernetes and use moby
as the container engine, we are forcing the factory reset first.

In addition to this, we check if Rancher Desktop was installed manually
with `brew` command. If this is true we uninstall that package to avoid
any conflict.