[DT-142] ih-pre-commit integration (#40)
This adds a new, custom pre-commit handler (ih-pre-commit) which resides
in the image-builder repository. This change updates the git step to
remove its installation of the standard pre-commit hook, and updates the
toolrepos step to install the ih-pre-commit step as a template and to
update all existing IH repositories.

This also adds a couple useful tools as dependency, yq and ripgrep.