DT-158 permission for file copy  and architecture detected (#53)
In order to automatically configure Rancher Desktop, a file should be
copied into the computer systems of the users. Sometimes this requires
additional permissions. So I added the sudo command.

In addition to this we are detecting the architecture of the computer to
provide the brew version that will be executed.

In addition to this, we are checking if Rancher Desktop was installed
successfully in order to continue with the settings such as symlinks.

---------

Co-authored-by: SteveRuble <SteveRuble@users.noreply.github.com>