fix: reduce impact of multiple brew archs
If a user has x86 brew AND arm64 brew installed on an M1 they can get weird problems that are hard to understand because which is active at a given time can vary during the execution of ih-setup or other scripts. This change reduces the chances of different versions of brew becoming active at different times.