[DP-2769] Update default Go to 1.21 for forward compatiability (#63)
**Summary**
Recently golang released version 1.21 which allows go.mod files to
specify the golang version to use and if not available, download it.
This will allow services that use image-builder to not have to specify
an asdf golang version and instead be able to use the golang version
installed in image-builder. In order to use this functionality however,
the default version of golang in `~/.tool-versions` needs to be a
minimum 1.21.