Change "Your username" to "Your email username"
For context; when running this script I was unsure if the username it asked for was referring to my GitHub username (which I had just entered), or some other credential. Haven't heard the first chunk of an email address referred to as "username", so adding some more words for clarity.