CLOUD-4800: fix mac m2+ check to work with m4 macs (#108)
Related to: https://github.com/ConsultingMD/homebrew-ih-public/pull/90

The current logic is only expecting the issue with m2 and m3 macs (not
m1 macs), but we also want to use Rancher's VZ and virtuofs emulation on
m4 macs. This PR improves the current logic to work with more macs to
fix potential issues with mounting volumes.