DP-4818: Try and fast-forward if not on default branch (#123)
Our auto update script will update your repo only if you are on the
default branch.

Sometimes these repos won't be on the default branch. So try our best to
pull the latest into the current branch using a fast-forward, and abort
if not. We also autostash changes if they exist.

This also adds some more development repos to the list of auto-updated
repos.